mem: Parse mem=xxx command line correctly
Fix the data type to parse the argument for the main memory correctly.

+++ sys/arch/i386/include/l4/linux_compat.h
@@ -36,7 +36,7 @@ int bitmap_find_free_region(unsigned long *bitmap, int bits, int order);
 /* XXX This is most probably wrong */
 #define THREAD_SIZE		(2 * PAGE_SIZE)
 
-unsigned long long memparse(const char *ptr, char **retptr);
+unsigned long memparse(const char *ptr, char **retptr);
 
 /* form include/kernel.h */
